test: migrate to Vitest
