+++ eslint.config.mjs
@@ -37,9 +37,7 @@ export default [
     languageOptions: {
       globals: {
         ...globals.node,
-        ...globals.jest,
       },
-      sourceType: 'commonjs',
     },
   },
   {
