+++ package.json
@@ -33,7 +33,7 @@
     "prettier": "prettier --check .",
     "prettier-write": "prettier --write .",
     "test": "npm run build && npm run test-only && npm run eslint && npm run prettier",
-    "test-only": "jest",
+    "test-only": "vitest run",
     "vite": "vite --open",
     "vite-build": "vite build",
     "vite-preview": "vite preview"
@@ -77,7 +77,6 @@
   },
   "homepage": "https://github.com/cheminfo/openchemlib-js",
   "devDependencies": {
-    "@types/jest": "^29.5.14",
     "@types/node": "^22.13.1",
     "autoprefixer": "^10.4.20",
     "benchmark": "^2.1.4",
@@ -88,13 +87,13 @@
     "globals": "^15.14.0",
     "gwt-api-exporter": "^2.0.0",
     "image-js": "0.0.0-next-692b15513a",
-    "jest": "^29.7.0",
     "openchemlib-utils": "^6.5.2",
     "postcss": "^8.5.1",
     "prettier": "^3.4.2",
     "tailwindcss": "^3.4.13",
     "typescript": "^5.7.3",
     "vite": "^6.0.11",
+    "vitest": "^3.0.5",
     "yargs": "^17.7.2"
   },
   "volta": {
